backfastapi/requirements.txt
uvicorn
subliminal
pydantic
python-dotenv
google-generativeai
fastapi
chardet